[Get_by_id] Refactored priority and brandId to be integers instead of longs. All tests passing calling to getbyid

+++ src/main/java/com/inditex/hiring/application/offer/AddOfferCommand.java
@@ -10,11 +10,11 @@ import java.time.Instant;
 @Value
 @Builder
 public class AddOfferCommand implements Command {
-    long brandId;
+    int brandId;
     long priceListId;
     Instant startDate;
     Instant endDate;
-    long priority;
+    int priority;
     String partNumber;
     BigDecimal price;
     String currencyISO;
