+++ src/test/java/com/inditex/hiring/e2e/offer/models/response/AddOfferResponse.java
@@ -1,5 +1,5 @@
 package com.inditex.hiring.e2e.offer.models.response;
 
-public record AddOfferResponse(long id, long brandId, long priceListId, String startDate,
-                               String endDate, long priority, String partNumber, double price, String currencyISO) {
+public record AddOfferResponse(long id, int brandId, long priceListId, String startDate,
+                               String endDate, int priority, String partNumber, double price, String currencyISO) {
 }
